+++ mission-4/src/services/commentService.ts
@@ -1,8 +1,24 @@
+import type {
+  CommentParams,
+  GetCommentList,
+  PatchComment,
+  PostComment,
+} from '@/types/comment.typs.js';
 import prisma from '../lib/prisma.js';
+import type { Prisma } from '@prisma/client';
+import { ForbiddenError } from '@/lib/errors.js';
 
-async function getCommentListService({ cursorId, pageSize, parentType }) {
-  const where = parentType === 'product' ? { productId: { not: null } } : { articleId: { not: null } };
-  const query = {
+async function authorization({ userId, commentId }: CommentParams): Promise<boolean> {
+  const comment = await prisma.comment.findUniqueOrThrow({
+    where: { id: commentId },
+  });
+  return comment.userId === userId;
+}
+
+async function getCommentListService({ cursorId, pageSize, parentType }: GetCommentList) {
+  const where =
+    parentType === 'products' ? { productId: { not: null } } : { articleId: { not: null } };
+  const query: Prisma.CommentFindManyArgs = {
     where,
     select: {
       id: true,
@@ -19,42 +35,51 @@ async function getCommentListService({ cursorId, pageSize, parentType }) {
     take: pageSize,
   };
   if (cursorId) {
-    query.cursor = { id: cursorId };
-    query.skip = 1;
+    query['cursor'] = { id: cursorId };
+    query['skip'] = 1;
   }
   const comment = await prisma.comment.findMany(query);
   return comment;
 }
 
-async function postCommentService({ userId, parentId, parentType, content }) {
+async function postCommentService({ userId, parentId, parentType, content }: PostComment) {
+  const singularParentType = parentType.endsWith('s') ? parentType.slice(0, -1) : parentType;
   const comment = await prisma.comment.create({
     data: {
       content,
-      [`${parentType}Id`]: parentId,
+      [`${singularParentType}Id`]: parentId,
       userId,
     },
     include: {
-      [parentType]: true,
+      [singularParentType]: true,
     },
   });
   return comment;
 }
 
-async function patchCommentService({ id, content }) {
-  const comment = await prisma.comment.update({
-    where: id,
-    data: {
-      content,
-    },
-  });
-  return comment;
+async function patchCommentService({ commentId, userId, content }: PatchComment) {
+  if (await authorization({ userId, commentId })) {
+    const comment = await prisma.comment.update({
+      where: { id: commentId },
+      data: {
+        content,
+      },
+    });
+    return comment;
+  } else {
+    throw new ForbiddenError('수정 권한이 없습니다.');
+  }
 }
 
-async function deleteCommentService({ id }) {
-  const comment = await prisma.comment.delete({
-    where: id,
-  });
-  return comment;
+async function deleteCommentService({ commentId, userId }: CommentParams) {
+  if (await authorization({ userId, commentId })) {
+    const comment = await prisma.comment.delete({
+      where: { id: commentId },
+    });
+    return comment;
+  } else {
+    throw new ForbiddenError('삭제 권한이 없습니다.');
+  }
 }
 
 export { getCommentListService, postCommentService, patchCommentService, deleteCommentService };
